+++ bindings/Python/README.md
@@ -4,7 +4,7 @@
 ## Quick examples using Python:
 
 ```python
->>> from symspell_pybind import SymspellPy
+>>> from symspell_rs import SymspellPy
 >>> sym_spell = SymspellPy(max_distance=2,prefix_length=7,count_threshold=1)
 >>> if not sym_spell.load_dictionary("./data/frequency_dictionary_en_82_765.txt",0,1," "):
       print("File Not Found")
