--- bindings/Python/symspell_pybind/__init__.py
@@ -1,1 +1,0 @@
-from .symspell_pybind import *
