+++ bindings/Python/symspell_rs/__init__.py
@@ -1,0 +1,1 @@
+from .symspell_rs import *
